requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_groq
fastapi
unicorn
langserve[all]
pydantic
langchain_community
langgraph
streamlit
tavily-python
